workspace-kit: Separate rings.go into ringsN.go
